refactor: move options and add dry run
